+++ GameTest/src/movement/MoveDownRight.java
@@ -5,7 +5,7 @@ public class MoveDownRight extends Thread{
 	public void run() {
 		try {
 			while (SimpleDungeonCrawler.movingRight && SimpleDungeonCrawler.movingDown) {
-				int speed = SimpleDungeonCrawler.diagSpeed;
+				double speed = SimpleDungeonCrawler.diagSpeed;
 				SimpleDungeonCrawler.movePlayer(speed, 0);
 				SimpleDungeonCrawler.movePlayer(0, speed);
 				Thread.sleep(SimpleDungeonCrawler.refreshRate);
