+++ GameTest/src/movement/MoveRight.java
@@ -8,7 +8,7 @@ public class MoveRight extends Thread {
 		try {
 			while (SimpleDungeonCrawler.movingRight) {
 				if (!SimpleDungeonCrawler.movingDown && !SimpleDungeonCrawler.movingUp) {
-					int speed = SimpleDungeonCrawler.playerSpeed;
+					double speed = SimpleDungeonCrawler.playerSpeed;
 					SimpleDungeonCrawler.movePlayer(speed, 0);
 					count++;
 					if (count % 10 < 5) {
