+++ GameTest/src/movement/MoveUp.java
@@ -6,7 +6,7 @@ public class MoveUp extends Thread {
 		try {
 			while (SimpleDungeonCrawler.movingUp) {
 				if (!SimpleDungeonCrawler.movingRight && !SimpleDungeonCrawler.movingLeft) {
-					int speed = SimpleDungeonCrawler.playerSpeed;
+					double speed = SimpleDungeonCrawler.playerSpeed;
 					SimpleDungeonCrawler.movePlayer(0, -speed);
 				}					
 				Thread.sleep(SimpleDungeonCrawler.refreshRate);
